+++ src/aux-common/bots/Bot.ts
@@ -1,3 +1,5 @@
+import { TagEditOp } from '../aux-format-2';
+
 export type PartialBot = Partial<Bot>;
 
 export type AuxDomain = 'builder' | 'player';
@@ -24,6 +26,16 @@ export const GET_TAG_MASK_SYMBOL = Symbol('get_tag_mask');
  * Defines a symbol that is used to get all the tag masks on a runtime bot.
  */
 export const CLEAR_TAG_MASKS_SYMBOL = Symbol('clear_tag_masks');
+
+/**
+ * Defines a symbol that is used to edit a tag.
+ */
+export const EDIT_TAG_SYMBOL = Symbol('edit_tag');
+
+/**
+ * Defines a symbol that is used to edit a  tag mask.
+ */
+export const EDIT_TAG_MASK_SYMBOL = Symbol('edit_tag_mask');
 
 /**
  * Defines an interface for a bot in a script/formula.
@@ -89,6 +101,20 @@ export interface RuntimeBot {
      * @param space The space that the masks should be cleared from. If not specified then all tag masks in all spaces will be cleared.
      */
     [CLEAR_TAG_MASKS_SYMBOL]: (space?: string) => any;
+
+    /**
+     * A function that can manipulate a tag using the given edit operations.
+     */
+    [EDIT_TAG_SYMBOL]: (tag: string, ops: TagEditOp[]) => any;
+
+    /**
+     * A function that can manipulate a tag mask using the given edit operations.
+     */
+    [EDIT_TAG_MASK_SYMBOL]: (
+        tag: string,
+        ops: TagEditOp[],
+        space: string
+    ) => any;
 }
 
 /**